requirements.txt
fastapi==0.116.1
httpx==0.28.1
pydantic==2.13.0
PyMuPDF==1.26.3
python-multipart==0.0.20
PyYAML==6.0.2
uvicorn==0.35.0
pytest==8.4.1
